refactor(sparkline): 调整了类型定义
