+++ src/ui/sparkline/types.ts
@@ -1,4 +1,21 @@
+import { Linear, Band } from '@antv/scale';
 import { ShapeAttrs, ShapeCfg } from '../../types';
+
+export type Point = [number, number];
+export type Line = Point[];
+export type Data = number[][];
+export type Scales = {
+  y: Linear;
+} & (
+  | {
+      type: 'line';
+      x: Linear;
+    }
+  | {
+      type: 'column';
+      x: Band;
+    }
+);
 
 export type SparklineOptions = ShapeCfg & {
   data?: number[] | number[][];
